2.1.0

+++ SoraPassport/Common/Model/Chain.swift
@@ -6,7 +6,7 @@
 import Foundation
 import IrohaCrypto
 
-typealias SNAddressType = UInt8
+typealias SNAddressType = UInt16
 
 enum Chain: String, Codable, CaseIterable {
     case polkadot = "Polkadot"
